+++ railties/lib/rails/application/configuration.rb
@@ -398,6 +398,10 @@ module Rails
             active_job.enqueue_after_transaction_commit = true
           end
 
+          if respond_to?(:action_text)
+            action_text.editor = :lexxy
+          end
+
           ActiveSupport.raise_on_invalid_time_zone_parse = true
         else
           raise "Unknown version #{target_version.to_s.inspect}"